move room create popup into own component
